ensure vector search does not prefetch opensearch results
